requirements.txt
pyperclip==1.8.2
pywin32==306
pystray==0.19.4
Pillow==10.0.1
keyboard==1.13.1
